update to compose_dataset.py and resulting wiki csv's
vanilla_wiki_data has csv's with only a list of the evidence sentences (no links, etc). test_, eval_, and train_data are separated, with test_ and eval_ having even distributions of supports, refutes, and nei.

in the csv's, 'verifiable' is a binary: 1 for yes and 0 for no.

'label' is also numeric, with the following legend:
NEI: 0
Supports: 1
Refutes: 2